# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Takeaway App API",
        "description": "API documentation for the Takeaway App",
        "contact": {
            "email": "linwenhao5@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get all addresses for the authenticated customer",
                "operationId": "9c9db908a39758eb6b4157138546d088",
                "responses": {
                    "200": {
                        "description": "List of addresses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "addresses": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "张三"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "0612345678"
                                                    },
                                                    "street": {
                                                        "type": "string",
                                                        "example": "Wagenweg"
                                                    },
                                                    "house_number": {
                                                        "type": "string",
                                                        "example": "12B"
                                                    },
                                                    "postcode": {
                                                        "type": "string",
                                                        "example": "1442CE"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Purmerend"
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "Netherlands"
                                                    },
                                                    "customer_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-01T00:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-01T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Failed to get addresses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to get addresses"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Exception message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Add a new address for the authenticated customer",
                "operationId": "0641a0f33b3bb27ab067b144573b0018",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "street",
                                    "house_number",
                                    "postcode",
                                    "city",
                                    "country"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Recipient name",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "phone": {
                                        "description": "Dutch phone number",
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "Wagenweg"
                                    },
                                    "house_number": {
                                        "type": "string",
                                        "example": "12B"
                                    },
                                    "postcode": {
                                        "type": "string",
                                        "example": "1442CE"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Purmerend"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Netherlands"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "address": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "张三"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "0612345678"
                                                },
                                                "street": {
                                                    "type": "string",
                                                    "example": "Wagenweg"
                                                },
                                                "house_number": {
                                                    "type": "string",
                                                    "example": "12B"
                                                },
                                                "postcode": {
                                                    "type": "string",
                                                    "example": "1442CE"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Purmerend"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "Netherlands"
                                                },
                                                "customer_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-01T00:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed or business logic error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to add address",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to add address"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Exception message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token:",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}